questions load to dark whiteboard theme using alg. 4 public whiteboards have been setup. when they are loaded they wipe all data from previous sessions and fill with welcome.
